### 2020-10-24 #2779 @whedon
[PRE REVIEW]: Synthia: multi-dimensional synthetic data generation in Python
**Submitting author:** @dmey (<a href="http://orcid.org/0000-0002-7071-7547">D. Meyer</a>)
**Repository:** <a href="https://github.com/dmey/synthia" target ="_blank">https://github.com/dmey/synthia</a>
**Version:** 1.0.0
