### 2020-10-24 #2779 @kyleniemeyer
Hi @dmey, thanks for your submission to JOSS. Due to the relatively small size of your software package, the editorial board is going to take a closer look at whether it falls within our scope.
