### 2020-10-24 #2779 @dmey
Hi @kyleniemeyer, many thanks for letting me know. In case this may be of relevance to the board, this package have been used in two papers (currently in preparation) which I am planning to submit in the next few weeks. Furthermore, the tool is novel in its approach and is likely to be cited in future machine learning (ML) groups.
