### 2020-10-24 ac7efd7
remove obsolete comments
