+++ setup.py
@@ -8,7 +8,7 @@ with open(os.path.join(this_dir, "README.md"), encoding='utf-8') as f:
 
 setup(
     name='synthia',
-    version='0.1.0',
+    version='0.1.1',
     description='Multi-dimensional synthetic data generation in Python',
     long_description=long_description,
     long_description_content_type="text/markdown",
