### 2020-10-24 #2779 @dmey
@kyleniemeyer just as clarification to my previous message -- as I am going to upload the scientific papers that make use of/cite Synthia on arXiv in a couple of weeks while their peer-review takes place, I can update this thread with links to those respective papers. Originally, I thought that this was going to be discussed during review but I am more than happy to wait here if that will make it easier to show the novelty and contribution of this tool to the community.
