### 2020-10-26 #2779 @VivianePons
I'm having a look at the paper regarding the scope query requested by @kyleniemeyer . Is it normal that the paper is extremely short? The Github pdf only contains a "Summary" and "Acknowledgments" sections, it seems rather incomplete and I wonder if this is an involuntary mistake
