### 2020-10-26 #2779 @dmey
@VivianePons thanks for looking into this. My understanding is that the summary paper needs to be very short -- abstract like -- as it is meant only as summary of the motivation and purpose of the tool and because the purpose of the review is to review the software rather than paper as done in more traditional journals. I have checked again at https://joss.readthedocs.io/en/latest/submitting.html and it says that the summary paper should be between 250-1000 words but I am more than happy to extend this, especially given that my first draft was much much longer and cut it down considerably at submission to make it more to the point.
