### 2020-10-26 #2779 @VivianePons
Indeed, papers are rather short but they are still a bit more furnished. Look at our example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

In particular, papers should contain a "Statement of need" which is missing in your case. You can also have some other sections such as "Features", "Examples"

You can browse through our recent publications to give you an idea.
