### 2020-10-26 #2779 @VivianePons
In particular,  I would like to understand what your software adds specifically in terms of implementation. Considering the small amount of code, we might fear that it is mainly a python wrapper to some other tools like vinecopulib. Could you give us some information regarding this aspect?
