### 2020-10-29 #2779 @arfon
> Could you give us some information regarding this aspect?

@dmey - could you elaborate? This will help us making our editorial scope decision.
