### 2020-11-09 d8b7d59
add support for handling discrete quantities
