### 2020-11-10 7a70b93
move discrete ranks to end for vinecopulib
