### 2020-11-10 3261b51
Add support for handling discrete quantities (#9)
For vine copulas
