### 2020-11-10 badd67d
add seed kw arg to copula
