+++ tests/test_generators.py
@@ -38,6 +38,8 @@ def test_independent_dataset_generation():
     assert synthetic_data['a'].shape == (n_synthetic_samples, n_features[0])
     assert synthetic_data['b'].shape == (n_synthetic_samples, n_features[1])
 
+    np.testing.assert_equal(generator.generate(10, seed=42), generator.generate(10, seed=42))
+
 def test_independent_feature_generation_with_distribution():
     n_samples = 20
     n_features = 2
@@ -82,6 +84,8 @@ def test_copula_dataset_generation():
 
     assert synthetic_data['a'].shape == (n_synthetic_samples, n_features[0])
     assert synthetic_data['b'].shape == (n_synthetic_samples, n_features[1])
+
+    np.testing.assert_equal(generator.generate(10, seed=42), generator.generate(10, seed=42))
 
 def test_fpca_dataset_generation():
     n_samples = 200
@@ -143,6 +147,8 @@ def test_vine_copula_feature_generation():
 
     assert synthetic_data.shape == (n_synthetic_samples, n_features)
     assert isinstance(synthetic_data, xr.DataArray)
+
+    np.testing.assert_equal(generator.generate(10, seed=42), generator.generate(10, seed=42))
 
 def test_copula_ndarray_feature_generation():
     n_samples = 200
