### 2020-11-10 f23aa21
Add seed kw arg to copula classes (#12)
closes #11.
