+++ synthia/copulas/vine.py
@@ -58,7 +58,7 @@ class VineCopula(Copula):
         var_types = ["d" if d else "c" for d in is_discrete]
         self.model = pv.Vinecop(rank_standardized, controls=self.controls, var_types=var_types)
 
-    def generate(self, n_samples: int, qrng=False, num_threads=1) -> np.ndarray:
+    def generate(self, n_samples: int, qrng: bool=False, num_threads: int=1, seed: Optional[int]=None) -> np.ndarray:
         """Generate n_samples Vine copula entries.
 
         Args:
@@ -67,7 +67,8 @@ class VineCopula(Copula):
         Returns:
             2D array of shape (sample, feature) with Vine copula entries.
         """
-        u_sim = self.model.simulate(n_samples, qrng=qrng, num_threads=num_threads)
+        seeds = [] if seed is None else [seed]
+        u_sim = self.model.simulate(n_samples, qrng=qrng, num_threads=num_threads, seeds=seeds)
         return u_sim
 
     # Pickling support
