+++ CHANGELOG.txt
@@ -1,0 +1,9 @@
+0.2.0
+- Add support for handling discrete quantities (#9).
+- Add support for setting a seed when generating new samples (#11).
+
+0.1.1
+- Fix qrng argument for pyvinecopulib due to vinecopulib/pyvinecopulib#68 and vinecopulib/pyvinecopulib#69.
+
+0.1.0
+- First public release.
