+++ setup.py
@@ -16,8 +16,8 @@ setup(
     license="MIT",
     author='D. Meyer, T. Nagler',
     classifiers=[
-        "Development Status :: 3 - Alpha",
-        "Intended Audience :: Developers",
+        "Development Status :: 4 - Beta",
+        "Intended Audience :: Science/Research",
         "Programming Language :: Python :: 3",
     ],
     packages=find_packages(),
