+++ synthia/util.py
@@ -101,7 +101,7 @@ def to_unstacked_dataset(arr: np.ndarray, stack_info: StackInfo) -> xr.Dataset:
     return ds
 
 def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset],
-                     types: Optional[Union[str, Dict[str,str], Dict[int,str]]]) -> Tuple[xr.DataArray, dict]:
+                     types: Optional[Union[str, Dict[str,str], Dict[int,str]]]=None) -> Tuple[xr.DataArray, dict]:
     data_info = {}
     if isinstance(data, xr.Dataset):
         data, stack_info = to_stacked_array(data)
@@ -126,7 +126,7 @@ def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset],
     features = []
     for i in range(data.shape[1]):
         if types_[i] == 'cat':
-            one_hot, categories[i] = categories_to_one_hot(data[:,i])
+            one_hot, categories[i] = categories_to_one_hot(data[:,i].values)
             features.append(one_hot)
         else:
             features.append(data[:,i:i+1])
@@ -137,22 +137,26 @@ def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset],
 
 def categories_to_one_hot(data: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
     assert data.ndim == 1
+    # determine all categories occurring in the data
     categories = np.unique(data)
     assert categories.ndim == 1
-    indices = np.empty(data.shape, np.uint16)
+    n_categories = len(categories)
+    # re-index categories, starting from 0
+    indices_dtype = None
+    for dtype in [np.uint8, np.uint16, np.uint32, np.uint64]:
+        if n_categories < np.iinfo(dtype).max:
+            indices_dtype = dtype
+    assert indices_dtype
+    indices = np.empty(data.shape, indices_dtype)
     for index, cat_val in enumerate(categories):
         indices[data == cat_val] = index
-    one_hot = indices_to_one_hot(indices, len(categories))
-    assert one_hot.shape == (len(data), len(categories))
+    # create one-hot vectors
+    one_hot = np.eye(n_categories)[indices]    
+    assert one_hot.shape == (len(data), n_categories)
     return one_hot, categories
 
-# https://stackoverflow.com/a/42874900
-def indices_to_one_hot(data, nb_classes):
-    """Convert an iterable of indices to one-hot encoded labels."""
-    targets = np.array(data).reshape(-1)
-    return np.eye(nb_classes)[targets]
-
 def one_hot_to_categories(data: np.ndarray, categories: np.ndarray) -> np.ndarray:
+    assert isinstance(data, np.ndarray)
     assert data.ndim == 2
     assert categories.ndim == 1
     assert data.shape[1] == len(categories)
@@ -164,9 +168,21 @@ def one_hot_to_categories(data: np.ndarray, categories: np.ndarray) -> np.ndarra
 def from_feature_array(data: np.ndarray, data_info: dict) -> Union[np.ndarray, xr.DataArray, xr.Dataset]:
     categories = data_info.get('categories')
     if categories:
-        # Map one-hot encoded categories into original values.
-        # TODO
-        pass
+        # Transform one-hot encoded categories into original values.
+        features = []
+        i_onehot = 0
+        i_normal = 0
+        while i_onehot < data.shape[1]:
+            if i_normal in categories:
+                n_categories = len(categories[i_normal])
+                feature = one_hot_to_categories(data[:, i_onehot:i_onehot+n_categories], categories[i_normal])
+                features.append(feature.reshape(-1, 1))
+                i_onehot += n_categories
+            else:
+                features.append(data[:, i_onehot:i_onehot+1])
+                i_onehot += 1
+            i_normal += 1
[…]
