### 2020-11-12 676c714
update installation instrictions
