+++ README.md
@@ -1,15 +1,20 @@
 <div align="center">
-  <img src="assets/img/logo.png" alt="synthia" height="120"><br><br>
+  <img src="assets/img/logo.png" alt="synthia" height="120">
+  ---
+</div>
 
+<div align="center">
 
-  [Overview](#overview) | [How to cite](#how-to-cite) | [Contributing](#contributing) | [Development notes](#development-notes) | [Copyright and license](#copyright-and-license) | [Acknowledgements](#acknowledgements)
+  ![PyPI](https://img.shields.io/pypi/v/synthia) ![CI](https://github.com/dmey/synthia/workflows/CI/badge.svg)
+
+  [Overview](#overview) | [Documentation](#documentation) | [How to cite](#how-to-cite) | [Contributing](#contributing) | [Development notes](#development-notes) | [Copyright and license](#copyright-and-license) | [Acknowledgements](#acknowledgements)
 </div>
 
 ## Overview
 
-Synthia is a tool for generating (continuous) multi-dimensional data in Python. It has a simple and succinct API to natively handle multidimensional data using [xarray](https://xarray.pydata.org)'s labeled arrays and datasets. For example, given some vertical profiles of atmospheric temperature, we can use Synthia to generate new but statistically similar profiles.
+Synthia is a tool for generating multi-dimensional data in Python. It has a simple and succinct API to natively handle multidimensional data using [xarray](https://xarray.pydata.org)'s labeled arrays and datasets. It supports different methods of data generation such as[ functional Principle Component Analysis (fPCA)](https://dmey.github.io/synthia/fpca.html), and Parametric (Gaussian) and Vine [copula models](https://dmey.github.io/synthia/copula.html). Depending on the type of model, it can be used to fit and generate different types of multivariate data with (or a mix of) discrete, categorical, and continuous variables. For example, given some vertical profiles of atmospheric temperature, we can use Synthia to generate new but statistically similar profiles using a copula and fPCA model in three lines of code (Table 1). For more information about the features included in the latest version of Synthia, please see the [Synthia website](https://dmey.github.io/synthia).
 
-**Example application of Gaussian and fPCA classes in Synthia used to generate random profiles of atmospheric temperature similar to those included in the SAF dataset (REF). The xarray dataset structure is maintained and returned by Synthia.**
+**Table 1. Example application of Gaussian and fPCA classes in Synthia used to generate random profiles of atmospheric temperature similar to those included in the SAF dataset (REF). The xarray dataset structure is maintained and returned by Synthia.**
 
 | Source                                   | Synthetic with Gaussian Copula                     | Synthetic with fPCA                            |
 | ---------------------------------------- | -------------------------------------------------- | ---------------------------------------------- |
@@ -20,22 +25,30 @@ Synthia is a tool for generating (continuous) multi-dimensional data in Python. 
 | ![dd](./assets/img/temperature_true.png) | ![dd](./assets/img/temperature_synth_gaussian.png) | ![dd](./assets/img/temperature_synth_fPCA.png) |
 
 
-For installation instructions, tutorials and general information, see https://dmey.github.io/synthia.
+## Documentation
+
+For installation instructions, getting started guides and tutorials, background information, and API reference summaries, please see the [Synthia website](https://dmey.github.io/synthia).
 
 
 ## How to cite
 
-[TODO]
+Please cite the software summary paper and software version using the following Digital Object Identifiers (DOIs) to [generate citations in your preferred style](https://citation.crosscite.org/):
+
+| Software summary paper | Software version* |
+| ---------------------- | ----------------- |
+| [TODO](TODO)           | [TODO](TODO)      |
+
[…]
