### 2020-11-12 792ebc1
Support categorical data for pyvinecopulib (#13) (closes #10)
