+++ synthia/util.py
@@ -100,7 +100,8 @@ def to_unstacked_dataset(arr: np.ndarray, stack_info: StackInfo) -> xr.Dataset:
     ds = xr.Dataset(unstacked)
     return ds
 
-def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset]) -> Tuple[xr.DataArray, dict]:
+def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset],
+                     types: Optional[Union[str, Dict[str,str], Dict[int,str]]]=None) -> Tuple[xr.DataArray, dict]:
     data_info = {}
     if isinstance(data, xr.Dataset):
         data, stack_info = to_stacked_array(data)
@@ -108,7 +109,6 @@ def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset]) -> Tuple
     else:
         if isinstance(data, xr.DataArray):
             data_info['da_info'] = dict(
-                coords=data.coords,
                 dims=data.dims,
                 name=data.name,
                 attrs=data.attrs
@@ -116,9 +116,74 @@ def to_feature_array(data: Union[np.ndarray, xr.DataArray, xr.Dataset]) -> Tuple
         data = xr.DataArray(data)
     assert data.ndim == 2, f'Input array must be 2D, given: {data.ndim}D'
     data_info['n_features'] = data.shape[1]
+
+    types_ = per_feature(types, data_info)
+    if not any(t == 'cat' for t in types_):
+        return data, data_info
+    
+    # Transform categorical features into one-hot vectors.
+    categories = {}
+    features = []
+    for i in range(data.shape[1]):
+        if types_[i] == 'cat':
+            one_hot, categories[i] = categories_to_one_hot(data[:,i].values)
+            features.append(one_hot)
+        else:
+            features.append(data[:,i:i+1])
+    data = xr.DataArray(np.concatenate(features, axis=1), dims=data.dims)
+    data_info['categories'] = categories
+
     return data, data_info
 
+def categories_to_one_hot(data: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+    assert data.ndim == 1
+    # determine all categories occurring in the data
+    categories = np.unique(data)
+    assert categories.ndim == 1
+    n_categories = len(categories)
+    # re-index categories, starting from 0
+    indices_dtype = None
+    for dtype in [np.uint8, np.uint16, np.uint32, np.uint64]:
+        if n_categories < np.iinfo(dtype).max:
+            indices_dtype = dtype
+    assert indices_dtype
+    indices = np.empty(data.shape, indices_dtype)
+    for index, cat_val in enumerate(categories):
+        indices[data == cat_val] = index
+    # create one-hot vectors
+    one_hot = np.eye(n_categories)[indices]    
+    assert one_hot.shape == (len(data), n_categories)
+    return one_hot, categories
+
+def one_hot_to_categories(data: np.ndarray, categories: np.ndarray) -> np.ndarray:
+    assert isinstance(data, np.ndarray)
+    assert data.ndim == 2
+    assert categories.ndim == 1
+    assert data.shape[1] == len(categories)
+    indices = np.argmax(data, axis=1)
+    out = categories[indices]
+    assert out.shape == (data.shape[0],)
+    return out
+
 def from_feature_array(data: np.ndarray, data_info: dict) -> Union[np.ndarray, xr.DataArray, xr.Dataset]:
+    categories = data_info.get('categories')
+    if categories:
+        # Transform one-hot encoded categories into original values.
+        features = []
+        i_onehot = 0
+        i_normal = 0
+        while i_onehot < data.shape[1]:
+            if i_normal in categories:
[…]
