### 2020-11-12 faab2b6
Prepare for 0.3.0 release
Update chnagelog and bump version to 0.3.0
