+++ docs/examples.rst
@@ -7,8 +7,11 @@ Generation
 .. toctree::
     :maxdepth: 1
 
+    examples/univariate
+    examples/multivariate-independent
     examples/multivariate-gaussian
     examples/fpca
+    examples/discrete-and-categorical
 
 Enhancement
 ***********
@@ -16,8 +19,3 @@ Enhancement
     :maxdepth: 1
 
     examples/enhancement
-
-Utilities
-*********
-.. toctree::
-    :maxdepth: 1
