### 2020-11-13 28cec64
Add more examples and improve docs (#14)
