+++ docs/citing.md
@@ -1,3 +1,9 @@
 # How to cite
 
-TODO
+Please cite the software summary paper and software version using the following Digital Object Identifiers (DOIs) to [generate citations in your preferred style](https://citation.crosscite.org/):
+
+| Software summary paper | Software version* |
+| ---------------------- | ----------------- |
+| *In Preparation*       | *In Preparation*  |
+
+*please make sure to cite the same version you are using with the correct DOI. For a list of all available versions see the list of available versions (In Preparation).
