### 2020-11-13 #2779 @dmey
@arfon @VivianePons @kyleniemeyer and @danielskatz many thanks for allowing me to get back to you this week. We have recently extended the documentation, added more examples, and reworded the paper to address what I think were your main concerns. We have also added a couple of new features, that is, the handling of discrete and categorical data in the last two releases which brings the number of lines of pure Python code to 1097 (please see `cloc` output below).

With regards to your individual questions --

> In particular, I would like to understand what your software adds specifically in terms of implementation.

@VivianePons thanks for raising this -- looking at the paper and repository with fresh eyes, I can see how this was unclear. I have now made changes to the repository, paper and website and hope that the changes make the purpose clearer. With regards to your specific question, Synthia can currently be used to model univariate and multivariate data, parameterize marginals with empirical and parametric methods and apply manipulations such as stretching and uniformization (I have added a summary at https://dmey.github.io/synthia/features.html). For multivariate data we support three different types of methods: fPCA, parametric (Gaussian) copula, and vine copula models and provide a pure Python implementation for the former two and rely on vinecopulib for the latter. Recently we have also added the capability to handle discrete and categorical data when using vine copulas.

>Considering the small amount of code, we might fear that it is mainly a python wrapper to some other tools like vinecopulib. Could you give us some information regarding this aspect?

We have tried to write Synthia succinctly and the current lines of pure Python code according to the `cloc` tool is 1097 (see below). The use of vinecopulib is important but it is not a required dependency. In our installation vinecopulib is also marked as an optional dependency (see https://dmey.github.io/synthia/installation.html). The amount of code that corresponds to the integration with vinecopulib is very very small, about 20-30 lines of code. Furthermore, although vinecopulib does play an important role in Synthia, its purpose is limited to the generation of vines not that of data generation in general.
As Synthia presents a new method for generation using multidimensional data in Python using fPCA, together with gaussian and vine copulas models, natively handle multidimensional arrays and datasets (essential in componential sciences), and the parametrizations and manipulation of univariate distribution in a single tool, I believe the paper is within scope.

The scope of the journal (https://joss.readthedocs.io/en/latest/submitting.html) indicates that [our **bold**]:

>JOSS publishes articles about research software. This definition includes software that: **solves complex modeling problems in a scientific context** (**physics, mathematics**, biology, medicine, social science, neuroscience, **engineering**); **supports the functioning of research instruments** or the **execution of research experiments**; **extracts knowledge from large data sets**; offers a mathematical library, or similar.

>JOSS publishes articles about software that represent **substantial scholarly effort on the part of the authors**. Your software should be a significant contribution to the available open source software that either **enables some new research challenges to be addressed** or **makes addressing research challenges significantly better** (e.g., faster, **easier**, **simpler**)

I cited a paper which is going to be submitted in the next few 10 days, I will let you know as soon as it's been deposited to that I can update the reference. And apologies for the long text but I thought it would be best to address everything in one long comment.

As a side note, I think there is a small issue with typesetting the figures in the paper (Table 1). Would it be possible to reduce the text size or change the width by a little so that the code blocks display as one liners. Otherwise I could move them to a different layout.

Output from the `cloc` command (local run, commit id: 0da044afc3c6d7bad0b60f54dcf21ba2fb6374be).
```
      54 text files.
      54 unique files.
      21 files ignored.

Language                     files          blank        comment           code
Python                          21            364            369           1097
Markdown                         9            117              0            206
YAML                             3             11              5             71
CSS                              1              7              7             61
INI                              1              0              0              2
HTML                             1              0              0              2
SUM:                            36            499            381           1439
```
