### 2020-11-16 #2779 @danielskatz
@openjournals/dev - any comments on this question from the author:
>As a side note, I think there is a small issue with typesetting the figures in the paper (Table 1). Would it be possible to reduce the text size or change the width by a little so that the code blocks display as one liners. Otherwise I could move them to a different layout.
