+++ .github/workflows/ci.yml
@@ -4,19 +4,26 @@ on: [push, pull_request]
 
 jobs:
   main:
-    runs-on: ubuntu-latest
-
+    name: ${{ matrix.os }} with python ${{ matrix.python-version }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
+        python-version: ['3.8', '3.9']
     steps:
     - uses: actions/checkout@v2
-    
-    - uses: goanpeca/setup-miniconda@v1
+
+    - uses: conda-incubator/setup-miniconda@v2
       with:
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
         activate-environment: synthia-test
         environment-file: environment-test.yml
 
     - name: Install Synthia (test env)
       # 'shell' required to activate environment.
-      # See https://github.com/goanpeca/setup-miniconda#important.
+      # See https://github.com/conda-incubator/setup-miniconda#IMPORTANT.
       shell: bash -l {0}
       run: pip install .
 
@@ -32,14 +39,16 @@ jobs:
         pytest tests/
 
     # Install all dev dependencies to build notebooks etc.
-    - uses: goanpeca/setup-miniconda@v1
+    - uses: conda-incubator/setup-miniconda@v2
       with:
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
         activate-environment: synthia
         environment-file: environment.yml
 
     - name: Install Synthia (dev env)
       # 'shell' required to activate environment.
-      # See https://github.com/goanpeca/setup-miniconda#important.
+      # See https://github.com/conda-incubator/setup-miniconda#IMPORTANT.
       shell: bash -l {0}
       run: pip install .
 
@@ -48,7 +57,7 @@ jobs:
       run: sphinx-build -v -b html docs/ docs/_build
 
     - name: Publish docs
-      if: github.event_name == 'push' && github.ref == 'refs/heads/master'
+      if: matrix.os == 'ubuntu-latest' && matrix.python-version == '3.8' && github.event_name == 'push' && github.ref == 'refs/heads/master'
       uses: peaceiris/actions-gh-pages@v3
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
