+++ environment-test.yml
@@ -1,6 +1,3 @@
 name: synthia-test
-channels:
-  - defaults
 dependencies:
-  - python=3.8 # when changing this, also change setup.py
   - pytest
