### 2020-11-18 c619bcc
seprate test and doc into two separate jobs
