+++ .github/workflows/ci.yml
@@ -10,6 +10,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
+        # when changing version, also change setup.py
         python-version: ['3.8', '3.9']
     steps:
     - uses: actions/checkout@v2
