### 2020-11-18 d1c1db2
hold off with python 3.9 until next pyvinecopulib release
