+++ .github/workflows/ci.yml
@@ -11,7 +11,7 @@ jobs:
       matrix:
         os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
         # when changing version, also change setup.py
-        python-version: ['3.8', '3.9']
+        python-version: ['3.8']
     steps:
     - uses: actions/checkout@v2
     - uses: conda-incubator/setup-miniconda@v2
