### 2020-11-18 9d5aeda
Enable CI on multiple OS and Python versions (#16)
