+++ environment.yml
@@ -3,7 +3,6 @@ channels:
   - defaults
   - conda-forge
 dependencies:
-  - python=3.8
   - pytest
   - numpy
   - scipy
