### 2020-11-24 #2779 @dmey
>@dmey can you give me some ideas for reviewers, please? 😊

@oliviaguest, sure -- from the list of [potential reviewers](https://bit.ly/joss-reviewers) I have identified the following people who may be suitable to review this submission:
```
@sbrugman, @Scivision, @terrytangyuan, @malmaud, @stsievert, @seba-1511, @yzhao062,
@NicolasHug, @glemaitre, @JesperDramsch, @zaxtax, @mnarayan, @justusschock
```

In case none is available please let me know and I can suggest more.
