### 2020-12-02 #2863 @khinsen
Question to @dmey, @tnagler, @letmaik: I am investigating authorship as required by JOSS reviewing guidelines. There is no doubt that @dmey is the main contributor to this project. Of the two other contributors, whose contributions look roughly equal in importance from the "size of commits" point of view, @tnagler is a co-author whereas @letmaik is merely acknowledged for "comments and contributions". If that's OK with all of you, it's fine with me as well of course, but it looks surprising.
