### 2020-12-02 #2863 @dmey
@khinsen it's good that you picked this up as it may appear a bit confusing -- all commits during development were squashed therefore the current metadata is not very indicative of the amount of work and contributions made during development. @tnagler made substantial contributions in the conceptual and practical development of the tool. @letmaik gave us suggestions and made contributions to the project.
