### 2020-12-02 #2863 @khinsen
Thanks @dmey and @letmaik! I realize that GitHub statistics share the problem of bibliometry in being superficial.
