### 2020-12-04 #2863 @oliviaguest
@khinsen the bot is just checking to see you are indeed doing the review, exactly like a reminder from a more "traditional" journal in your email inbox. Thus, you are able to ignore as you're already doing it.

If you are curious what commands @whedon does accept, I am sure you can figure that out too — haha: https://joss.readthedocs.io/en/latest/whedon.html
