### 2020-12-04 e8b0da3
clarify uncorrelatedness of principal components
See #18.
