+++ docs/fpca.md
@@ -31,6 +31,6 @@ This allows us to represent the high-dimensional vector $X$  by only a few numbe
 
 ## PCA for synthetic data generation
 
-PCA can be used to generate synthetic data for the high-dimensional vector $X$. For every instance $X_i$ in the data set, we compute the principal component scores $a_{i, 1}, \dots, a_{i, K}$. Because the principal components $v_1, \dots, v_K$ are orthogonal, the scores are necessarily uncorrelated and we may treat them as independent. We then fit a model $F_k$ for the marginal distribution of each score $a_k$. From these models, we can generate synthetic scores $\tilde a_k$ and transform them into a synthetic sample of $X$ via 
+PCA can be used to generate synthetic data for the high-dimensional vector $X$. For every instance $X_i$ in the data set, we compute the principal component scores $a_{i, 1}, \dots, a_{i, K}$. The scores are uncorrelated by construction and we treat them as independent. We then fit a model $F_k$ for the marginal distribution of each score $a_k$. From these models, we can generate synthetic scores $\tilde a_k$ and transform them into a synthetic sample of $X$ via 
 
 $$\tilde X = \mu + \sum_{k = 1}^K \tilde a_k v_{k}.$$
