### 2020-12-09 18e30b0
Explain how to run the tests after installation
