+++ docs/installation.md
@@ -10,6 +10,7 @@
 ## Optional dependencies
 
 - [pyvinecopulib](https://github.com/vinecopulib/pyvinecopulib): For vine copula and quasirandom number support.
+- [pytest](https://pytest.org): For running the test suite
 
 ## Instructions
 
@@ -21,4 +22,12 @@ or with optional dependencies
 
 ```
 pip install synthia[full]
+```
+
+After installation, you can run the test suite using
+
+```
+git clone https://github.com/dmey/synthia.git
+cd synthia
+python -m pytest -s tests/
 ```
