### 2020-12-09 #2863 @khinsen
@oliviaguest The checklist asks "Do references in the text use the proper citation syntax?", referring to a description of Markdown syntax for citations. What I have for review is a PDF file. I don't quite understand what I am supposed to do. Should I be looking at some Markdown source instead? If so, where can I find it?
