### 2020-12-09 #2863 @khinsen
@dmey After going through the checklist for the software, here comes my review of the paper itself. Overall, it looks very good, there is just one point I would like you to address:

"Do the authors describe how this software compares to other commonly-used packages?"

No. Please add a statement on this question, even if it is only "we are not aware of any other comparable software". The [Wikipedia page on Synthetic Data](https://en.wikipedia.org/wiki/Synthetic_data) lists one Open Source package ([DataGenerator](https://finraos.github.io/DataGenerator/)) and one publicly available package without a licence ([Dataset Generator](http://www.datasetgenerator.com/)), in addition to proprietary software.
