### 2020-12-09 #2863 @oliviaguest
@khinsen do you mean where is the paper being compiled from? Here: https://github.com/dmey/synthia/tree/joss-paper
