### 2020-12-09 #2863 @oliviaguest
@khinsen I thought (wrongly!) it was obvious when we ask for @whedon to do this:

> @whedon generate pdf from branch joss-paper

But either way — it's no problem at all and I am more than happy to explain it (there is so much to process for those who have never used JOSS before anyway). ☺️
