### 2020-12-11 #2863 @khinsen
@oliviaguest Everything is obvious if you do it often enough! Authors and editors are a lot more familiar with @whedon than reviewers, and for now that has been my only role in JOSS. We can usually do our jobs without ever talking to @whedon. Which I think is great, reviewing for JOSS is a real pleasure because of the absence of technical boilerplate tasks.
