### 2020-12-14 #2863 @oliviaguest
@khinsen Yes, reviewers do not *need* to interact with @whedon at all, indeed. But if you want to there is a list of commands you may run. In practice, it's basically just editors (and sometimes authors) who need to ask for the bot to do anything.

@mnarayan would you be able to give a rough ETA, please? 😊
