### 2020-12-24 #2863 @danielskatz
👋 @mnarayan - the original invitation to the JOSS organization may have expired - please use this link ☝️ to join, then you should be able to interact with the checklist.
