### 2021-01-09 #2863 @dmey
> @dmey After going through the checklist for the software, here comes my review of the paper itself. Overall, it looks very good, there is just one point I would like you to address:
>
> "Do the authors describe how this software compares to other commonly-used packages?"
>
> No. Please add a statement on this question, even if it is only "we are not aware of any other comparable software". The [Wikipedia page on Synthetic Data](https://en.wikipedia.org/wiki/Synthetic_data) lists one Open Source package ([DataGenerator](https://finraos.github.io/DataGenerator/)) and one publicly available package without a licence ([Dataset Generator](http://www.datasetgenerator.com/)), in addition to proprietary software.

@khinsen in the meantime I have updated the paper and added two recent data generation tools at the end of the second paragraph -- the ones you suggested do not appear to be maintained. I have also removed the example application as I think it did not add much to the paper (it's already on the website and readme) and were issues with the layout.
