### 2021-01-09 ef9336a
More coherent notation between the first two sections (#19)
