### 2021-01-09 d36aa9b
Temporarily limit sphinxcontrib-bibtex version to 1.x.x (#1137)
