+++ environment.yml
@@ -18,7 +18,7 @@ dependencies:
   - sphinx_rtd_theme
   - myst-parser
   - nbsphinx
-  - sphinxcontrib-bibtex
+  - sphinxcontrib-bibtex=1
   - sphinx-copybutton
   - pip:
     - pyvinecopulib
