+++ paper/paper.bib
@@ -1,0 +1,92 @@
+@book{Joe_2014,
+  title     = {Dependence {{Modeling}} with {{Copulas}}},
+  author    = {Joe, Harry},
+  year      = {2014},
+  edition   = {Zeroth},
+  publisher = {{Chapman and Hall/CRC}},
+  doi       = {10.1201/b17116},
+  isbn      = {978-0-429-10318-6},
+  language  = {en}
+}
+
+@article{Meyer2021,
+  doi       = {10.5194/gmd-2020-427},
+  url       = {https://doi.org/10.5194/gmd-2020-427},
+  year      = {2021},
+  publisher = {Copernicus {GmbH}},
+  author    = {David Meyer and Thomas Nagler and Robin J. Hogan},
+  title     = {Copula-based synthetic data generation for machine learning emulators
+		in weather and climate: application to a simple radiation model}
+}
+
+@inproceedings{Patki_2016,
+  title     = {The {{Synthetic Data Vault}}},
+  booktitle = {2016 {{IEEE International Conference}} on {{Data Science}} and {{Advanced Analytics}} ({{DSAA}})},
+  author    = {Patki, Neha and Wedge, Roy and Veeramachaneni, Kalyan},
+  year      = {2016},
+  pages     = {399--410},
+  publisher = {{IEEE}},
+  address   = {{Montreal, QC, Canada}},
+  doi       = {10.1109/DSAA.2016.49},
+  isbn      = {978-1-5090-5206-6}
+}
+
+@techreport{Eresmaa_2014,
+  author = {Eresmaa, Reima and McNally, Anthony},
+  year   = {2014},
+  month  = {10},
+  number = {NWPSAF-EC-TR-017},
+  pages  = {14},
+  title  = {Diverse profile datasets from the ECMWF 137-level short-range forecasts},
+  doi    = {10.13140/2.1.4476.8963}
+}
+
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {http://doi.org/10.5334/jors.148}
+}
+
+@book{McGraw-Hill-2003,
+  title     = {{{McGraw}}-{{Hill Dictionary}} of {{Scientific}} and {{Technical Terms}}},
+  author    = {{McGraw-Hill}},
+  year      = {2003},
+  edition   = {6th ed.},
+  publisher = {{McGraw-Hill Education}},
+  address   = {{New York}}
+}
+
+@software{nagler_thomas_2020_4287554,
+  author    = {Nagler, Thomas and
+                  Vatter, Thibault},
+  title     = {vinecopulib},
+  year      = 2020,
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.4287554},
+  url       = {https://doi.org/10.5281/zenodo.4287554}
+}
+
+@software{nagler_thomas_2020_4288292,
+  author    = {Nagler, Thomas and
+                  Vatter, Thibault},
+  title     = {pyvinecopulib},
+  year      = 2020,
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.4288292},
+  url       = {https://doi.org/10.5281/zenodo.4288292}
+}
+
+@misc{xu2018synthesizing,
+  title         = {Synthesizing Tabular Data using Generative Adversarial Networks},
+  author        = {Lei Xu and Kalyan Veeramachaneni},
+  year          = {2018},
+  eprint        = {1811.11264},
+  archiveprefix = {arXiv},
+  primaryclass  = {cs.LG}
+}
