+++ _sources/installation.md.txt
@@ -1,0 +1,24 @@
+# Installation
+
+## Required dependencies
+
+- Python (3.8 or later)
+- [numpy](http://www.numpy.org/)
+- [scipy](https://www.scipy.org/scipylib/index.html)
+- [xarray](http://xarray.pydata.org/)
+
+## Optional dependencies
+
+- [pyvinecopulib](https://github.com/vinecopulib/pyvinecopulib): For vine copula and quasirandom number support.
+
+## Instructions
+
+```
+pip install synthia
+```
+
+or with optional dependencies
+
+```
+pip install synthia[full]
+```
