+++ _sources/examples/univariate.ipynb.txt
@@ -1,0 +1,216 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Univariate"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Import libraries"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt\n",
+    "import synthia as syn"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Load sample data"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "ds_true = syn.util.load_dataset(name='SAF-Synthetic')\n",
+    "# We only select one level as we are dealing with univariate data\n",
+    "ds_true = ds_true.sel(level=-1)\n",
+    "ds_true['temperature_fl'].plot.hist()\n",
+    "plt.xlabel('Air temperature in K');"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Fit and generate using samples"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "generator = syn.CopulaDataGenerator()\n",
+    "# We fit and store the original samples and use them to recontuct new synthetic samples. \n",
+    "generator.fit(ds_true, copula=syn.IndependenceCopula())\n",
+    "N_SAMPLES = 1000\n",
+    "ds_synth_empirical = generator.generate(N_SAMPLES)\n",
+    "ds_synth_empirical['temperature_fl'].plot.hist()\n",
+    "plt.xlabel('Air temperature in K');"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Fit and generate using parametrized distribution"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {
+    "scrolled": false
+   },
+   "outputs": [],
+   "source": [
+    "generator = syn.CopulaDataGenerator()\n",
+    "# We now fit a parametric distribution instead of using the original samples (it may take some time)\n",
+    "dist_names = set(syn.DistributionParameterizer.get_dist_names())\n",
+    "# Remove all very slow distributions\n",
+    "dist_names -= set(['genexpon', 'levy_stable', 'recipinvgauss', 'vonmises', 'ncx2',\n",
+    "                   'kstwo', 'ncf', 'tukeylambda', 'nct', 'skewnorm', 'norminvgauss', 'gausshyper', 'geninvgauss'])\n",
+    "generator.fit(ds_true, copula=syn.IndependenceCopula(), \n",
+    "    parameterize_by=syn.DistributionParameterizer(dist_names, verbose=False))\n",
+    "ds_synth_param = generator.generate(1000)\n",
+    "ds_synth_param['temperature_fl'].plot.hist()\n",
+    "plt.xlabel('Air temperature in K');"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Stretching and unifomization"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
[…]
