+++ _sources/examples.rst.txt
@@ -1,0 +1,21 @@
+Examples
+========
+
+
+Generation
+**********
+.. toctree::
+    :maxdepth: 1
+
+    examples/univariate
+    examples/multivariate-independent
+    examples/multivariate-gaussian
+    examples/fpca
+    examples/discrete-and-categorical
+
+Enhancement
+***********
+.. toctree::
+    :maxdepth: 1
+
+    examples/enhancement

[…]
