### 2021-02-09 #2863 @mnarayan
@dmey This looks like a wonderful package. One that I might use myself in the near future!

I've opened a few issues with respect to the following
- [ ] **Comparisons to other packages could be expanded**. What are other packages people might use to simulate data (e.g. statsmodels comes to mind) and how is this different? Your package supports generating data for multivariate copula distributions and via fPCA. These are much more nonparametric compared to the typical parametric data generators that are typically available.
- [ ] **API clarification**. It would be helpful to add/explain what the different classes do `Data Generators`, `Parametrizer`, `Transformers`. I think generators and transformers are obvious but I only sort of understand `Parametrizers`. It is also confusing in the sense that people might think this has something to do with parametric distributions when you mean it to be something different. Is this API inspired by some convention elsewhere? If so it would be helpful to point to that. For instance, the generators are very similar to statsmodel generators.
- [ ] **Some expansion of usage & examples**. Consider my ask here optional. You have basic documentation, but I have questions as a potential user interested in simulating copula distributions that meet desired conditional dependencies.
