+++ synthia/parameterizers/const.py
@@ -6,6 +6,9 @@ import numpy as np
 from .parameterizer import Parameterizer
 
 class ConstParameterizer(Parameterizer):
+    """Preserves the size of the original data. No downsampling is performed
+    at fitting and no interpolation is perfomed at generation.
+    """
     def __init__(self, val) -> None:
         self.val = val
     
@@ -13,5 +16,7 @@ class ConstParameterizer(Parameterizer):
         pass
 
     def generate(self, n_samples: int) -> np.ndarray:
+        """Returns original samples without interpolation. 
+        """
         samples = np.full(n_samples, self.val)
         return samples
