+++ environment.yml
@@ -3,6 +3,7 @@ channels:
   - defaults
   - conda-forge
 dependencies:
+  - python=3.8
   - pytest
   - numpy
   - scipy
