### 2021-04-19 #2863 @oliviaguest
> Sorry for the delay in getting back, I think I have addressed all the issues that were previously raised. @khinsen and @mnarayan could you please let me know if you are OK with the changes I made to the repo and summary-paper?

Yes, please make sure you both have checked off everything in the list above. It seems only one or two things remain, and they are about the paper itself. Below this comment, I will ask whedon to compile the PDF again, so you can view the newest version.
