+++ paper/paper.bib
@@ -1,0 +1,184 @@
+@article{Dahmen_2019,
+  title = {{{SynSys}}: {{A Synthetic Data Generation System}} for {{Healthcare Applications}}},
+  shorttitle = {{{SynSys}}},
+  author = {Dahmen, Jessamyn and Cook, Diane},
+  year = {2019},
+  month = mar,
+  volume = {19},
+  pages = {1181},
+  issn = {1424-8220},
+  doi = {10.3390/s19051181},
+  journal = {Sensors},
+  language = {en},
+  number = {5}
+}
+
+@article{Harris_2020,
+  title = {Array Programming with {{NumPy}}},
+  author = {Harris, Charles R. and Millman, K. Jarrod and {van der Walt}, St{\'e}fan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and {van Kerkwijk}, Marten H. and Brett, Matthew and Haldane, Allan and {del R{\'i}o}, Jaime Fern{\'a}ndez and Wiebe, Mark and Peterson, Pearu and {G{\'e}rard-Marchant}, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
+  year = {2020},
+  month = sep,
+  volume = {585},
+  pages = {357--362},
+  issn = {0028-0836, 1476-4687},
+  doi = {10.1038/s41586-020-2649-2},
+  journal = {Nature},
+  language = {en},
+  number = {7825}
+}
+
+@article{Hoyer_2017,
+  title = {Xarray: {{N}}-{{D}} Labeled Arrays and Datasets in {{Python}}},
+  author = {Hoyer, S. and Hamman, J.},
+  year = {2017},
+  volume = {5},
+  doi = {10.5334/jors.148},
+  journal = {Journal of Open Research Software},
+  number = {1}
+}
+
+@book{Joe_2014,
+  title = {Dependence {{Modeling}} with {{Copulas}}},
+  author = {Joe, Harry},
+  year = {2014},
+  month = jun,
+  edition = {1st Edition},
+  publisher = {{Chapman and Hall/CRC}},
+  doi = {10.1201/b17116},
+  isbn = {978-0-429-10318-6},
+  language = {en}
+}
+
+@article{Kamthe_2021,
+  title = {Copula {{Flows}} for {{Synthetic Data Generation}}},
+  author = {Kamthe, Sanket and Assefa, Samuel and Deisenroth, Marc},
+  year = {2021},
+  month = jan,
+  archiveprefix = {arXiv},
+  eprint = {2101.00598},  
+  journal = {arXiv:2101.00598 [cs, stat]},
+  keywords = {Computer Science - Machine Learning,Statistics - Applications,Statistics - Machine Learning},
+  primaryclass = {cs, stat}
+}
+
+@inproceedings{Li_2020,
+  title = {{{SynC}}: {{A Copula}} Based {{Framework}} for {{Generating Synthetic Data}} from {{Aggregated Sources}}},
+  shorttitle = {{{SynC}}},
+  booktitle = {2020 {{International Conference}} on {{Data Mining Workshops}} ({{ICDMW}})},
+  author = {Li, Zheng and Zhao, Yue and Fu, Jialin},
+  year = {2020},
+  month = nov,
+  pages = {571--578},
+  publisher = {{IEEE}},
+  address = {{Sorrento, Italy}},
[…]
