### 2021-04-25 #2863 @oliviaguest
Oh... @dmey I thought those were all checked off the list(s) above. OK, @openjournals/joss-eics, please do not publish this just yet.

Can you @dmey please link me to what is left, I might have missed the issues, but ideally they should all be linked to from here for others (and me) to be aware of them.
