### 2021-04-26 #2863 @dmey
@mnarayan, when you have a moment can you please let me know if are happy with the resolution of the issues at https://github.com/openjournals/joss-reviews/issues/2863#issuecomment-776112028 and whether you are happy to close them.
