### 2021-04-26 #2863 @danielskatz
@dmey - I'm an AEiC, and I note that there was a message from whedon previously about there not being a Statement of Need section, which I think was not addressed. Perhaps you can split the summary into two sections, one of them being "Statement of Need"?
