### 2021-04-26 #2863 @dmey
@danielskatz thanks for letting me know about this. I thought I did addresses this by asking whether it was OK to blend the text in with the rest but you are right, I can see no mention of it above. In any case, would you be OK to keep it with the rest or do you require it to be a separate standalone section? I find that sometime this is not strongly enforced in other papers and personally, I find it to break the flow unnecessarily.
