### 2021-04-26 #2863 @danielskatz
At some point in the past (~6 months?), we made this a requirement for new submission, though we have been flexible in some cases where this is part of a section heading.
