### 2021-04-26 #2863 @dmey
> At some point in the past (~6 months?), we made this a requirement for new submissions, though we have been flexible in some cases where this is part of a section heading.

I see -- thanks for the clarification. If you think this is sufficiently clear and addressed in the current format, I would rather leave it as is given that the summary is already brief but please let me know if you think otherwise.
