### 2021-04-26 #2863 @dmey
> Just to be more clear, we do need a section that is titled (or includes the title) "Statement of Need." The two options are to break up the Summary section into two sections, or to change the title to "Summary and Statement of Need"

I can defiantly change the heading to "Summary and Statement of Need".

As a general comment and not really specific to this submission however I am not sure why this has to be emphasized so much in the heading. Surely a summary may contain the motivation for doing such and such as well as other aspects which do not need to be explicitly stated in the heading? But perhaps we can discuss this in a separately as it is more of a general meaning of what a statement of need should contain or be about!?
