### 2021-04-26 #2863 @danielskatz
Briefly, when working with thousands of submission, we have learned that the more explicit we can be and the more automated checks we have that are visible to both authors and editors, the easier the overall processing is, and the fewer repeated discussion we need to have
