+++ docs/citing.md
@@ -1,9 +1,29 @@
 # How to cite
 
-Please cite the software summary paper and software version using the following Digital Object Identifiers (DOIs) to [generate citations in your preferred style](https://citation.crosscite.org/):
+If you are using Synthia, please cite the following two papers using their respective Digital Object Identifiers (DOIs). Citations may be generated automatically using Crosscite's [DOI Citation Formatter](https://citation.crosscite.org/) or from the BibTeX entries below. If needed, you may also cite the specific software version with [its corresponding Zendo DOI](https://doi.org/10.5281/zenodo.4701278). 
 
-| Software summary paper | Software version* |
-| ---------------------- | ----------------- |
-| *In Preparation*       | *In Preparation*  |
+| Synthia Software                   | Software Application                                              |
+| ---------------------------------- | ----------------------------------------------------------------- |
+| DOI: 10.21105/joss.02863 | DOI: [10.5194/gmd-2020-427](https://doi.org/10.5194/gmd-2020-427) |
 
-*please make sure to cite the same version you are using with the correct DOI. For a list of all available versions see the list of available versions (In Preparation).
+```bibtex
+@article{Meyer_Nagler_2021,
+  title   = {Synthia: multidimensional synthetic data generation in Python},
+  author  = {David Meyer and Thomas Nagler},
+  year    = {2021},
+  doi     = {10.21105/joss.02863},
+  journal = {Journal of Open Source Software},
+  note    = {Under review}
+}
+
+@article{Meyer_Nagler_Hogan_2021,
+  title   = {Copula-Based Synthetic Data Generation for Machine Learning Emulators in Weather and Climate: Application to a Simple Radiation Model},
+  author  = {David Meyer and Thomas Nagler and Robin J. Hogan},
+  year    = {2021},
+  volume  = {2021},
+  doi     = {10.5194/gmd-2020-427},
+  journal = {Geoscientific Model Development Discussions},
+  note    = {Under review}
+}
+```
+
