+++ README.md
@@ -12,7 +12,7 @@ Synthetic data need to preserve the statistical properties of real data in terms
 
 Synthia is an open source Python package to model univariate and multivariate data, parameterize data using empirical and parametric methods, and manipulate marginal distributions. It is designed to enable scientists and practitioners to handle labelled multivariate data typical of computational sciences. For example, given some vertical profiles of atmospheric temperature, we can use Synthia to generate new but statistically similar profiles in just three lines of code (Table 1).
 
-Synthia supports three methods of multivariate data generation through fPCA, parametric (Gaussian) copula, and vine copula models for continuous (all), discrete (vine), and categorical (vine) variables. It has a simple and succinct API to natively handle [xarray](https://xarray.pydata.org)'s labelled arrays and datasets. It uses a pure Python implementation for fPCA and Gaussian copula, and relies on the fast and well tested C++ library [vinecopulib](https://github.com/vinecopulib/vinecopulib) through [pyvinecopulib](https://github.com/vinecopulib/pyvinecopulib)'s bindings for fast and efficient computation of vines. For more information about the latest version of Synthia, please see the [website](https://dmey.github.io/synthia).
+Synthia supports three methods of multivariate data generation through: (i) fPCA, (ii) parametric (Gaussian) copula, and (iii) vine copula models for continuous (all), discrete (vine), and categorical (vine) variables. It has a simple and succinct API to natively handle [xarray](https://xarray.pydata.org)'s labelled arrays and datasets. It uses a pure Python implementation for fPCA and Gaussian copula, and relies on the fast and well tested C++ library [vinecopulib](https://github.com/vinecopulib/vinecopulib) through [pyvinecopulib](https://github.com/vinecopulib/pyvinecopulib)'s bindings for fast and efficient computation of vines. For more information, please see the website at https://dmey.github.io/synthia.
 
 
 **Table 1**. *Example application of Gaussian and fPCA classes in Synthia. These are used to generate random profiles of atmospheric temperature similar to those included in the source data. The xarray dataset structure is maintained and returned by Synthia.*
@@ -33,35 +33,33 @@ For installation instructions, getting started guides and tutorials, background 
 
 ## How to cite
 
-When using Synthia, please cite the following two papers using the following Digital Object Identifiers (DOIs). You can generate citations in your preferred style with [DOI Citation Formatter](https://citation.crosscite.org/) or use the BibTeX entries below.
+If you are using Synthia, please cite the following two papers using their respective Digital Object Identifiers (DOIs). Citations may be generated automatically using Crosscite's [DOI Citation Formatter](https://citation.crosscite.org/) or from the BibTeX entries below. If needed, you may also cite the specific software version with [its corresponding Zendo DOI](https://doi.org/10.5281/zenodo.4701278). 
 
-| Application                                                  | Software            |
-| ------------------------------------------------------------ | ------------------- |
-| [10.5194/gmd-2020-427](https://doi.org/10.5194/gmd-2020-427) | 10.21105/joss.02863 |
+| Synthia Software                   | Software Application                                              |
+| ---------------------------------- | ----------------------------------------------------------------- |
+| DOI: 10.21105/joss.02863 | DOI: [10.5194/gmd-2020-427](https://doi.org/10.5194/gmd-2020-427) |
 
-```tex
-@article{Meyer2021,
-  doi       = {10.5194/gmd-2020-427},
-  year      = {2021},
-  publisher = {Copernicus {GmbH}},
[…]
