+++ paper/paper.bib
@@ -76,25 +76,24 @@
   language = {en}
 }
 
-@article{Meyer_2021,
-  title = {Machine {{Learning Emulation}} of {{3D Cloud Radiative Effects}}},
+@article{Meyer_2021b,
+  title = {Machine Learning Emulation of {{3D}} Cloud Radiative Effects},
   author = {Meyer, David and Hogan, Robin J. and Dueben, Peter D. and Mason, Shannon L.},
   year = {2021},
-  month = mar,
-  archiveprefix = {arXiv},
-  eprint = {2103.11919},
-  journal = {arXiv:2103.11919 [physics]},
-  keywords = {Computer Science - Machine Learning,Physics - Atmospheric and Oceanic Physics},
-  primaryclass = {physics}
+  journal = {Journal of Advances in Modeling Earth Systems},
+  doi = {10.1029/2021MS002550}
 }
 
 @article{Meyer_2021a,
-  title = {Copula-Based Synthetic Data Generation for Machine Learning Emulators in Weather and Climate: Application to a Simple Radiation Model},
-  author = {Meyer, David and Nagler, Thomas and Hogan, Robin J},
+  title = {Copula-Based Synthetic Data Augmentation for Machine-Learning Emulators},
+  author = {Meyer, David and Nagler, Thomas and Hogan, Robin J.},
   year = {2021},
-  volume = {2021},
-  doi = {10.5194/gmd-2020-427},
-  journal = {Geoscientific Model Development Discussions}
+  journal = {Geoscientific Model Development},
+  volume = {14},
+  number = {8},
+  pages = {5205--5215},
+  issn = {1991-9603},
+  doi = {10.5194/gmd-14-5205-2021}
 }
 
 @misc{Nagler_2020,
