+++ paper/paper.md
@@ -23,13 +23,13 @@ bibliography: paper.bib
 ---
 
 
-# Summary
+# Summary and Statement of Need
 
-Synthetic data -- artificially generated data that mimic the original (observed) data by preserving relationships between variables [@Nowok_2016] -- may be useful in several areas such as healthcare, finance, data science, and machine learning [@Dahmen_2019; @Kamthe_2021; @Nowok_2016; @Patki_2016]. As such, copula-based data generation models -- probabilistic models that allow for the statistical properties of observed data to be modelled in terms of individual behavior and (inter-)dependencies [@Joe_2014] -- have shown potential in finance, data science, and meteorology [@Kamthe_2021; @Li_2020; @Meyer_2021a; @Patki_2016]. Although copula-based data generation tools have been developed for tabular data -- e.g. the Synthetic Data Vault project using Gaussian copulas and generative adversarial networks [@Patki_2016; @Xu_2018], or the Synthetic Data Generation via Gaussian Copula [@Li_2020] -- in computational sciences such as weather and climate, data often consist of large, labelled multidimensional datasets with complex dependencies.
+Synthetic data -- artificially generated data that mimic the original (observed) data by preserving relationships between variables [@Nowok_2016] -- may be useful in several areas such as healthcare, finance, data science, and machine learning [@Dahmen_2019; @Kamthe_2021; @Nowok_2016; @Patki_2016]. As such, copula-based data generation models -- probabilistic models that allow for the statistical properties of observed data to be modelled in terms of individual behavior and (inter-)dependencies [@Joe_2014] -- have shown potential in finance, data science, and meteorology [@Kamthe_2021; @Li_2020; @Meyer_2021a; @Patki_2016]. Although copula-based data generation tools have been developed for tabular data -- e.g. the Synthetic Data Vault using Gaussian copulas and generative adversarial networks [@Patki_2016; @Xu_2018], or the Synthetic Data Generation via Gaussian Copula [@Li_2020] -- in computational sciences such as weather and climate, data often consist of large, labelled multidimensional datasets with complex dependencies.
 
-Here we introduce Synthia, an open-source multidimensional synthetic data generator written in Python for xarray's [@Hoyer_2017] labelled arrays and datasets with support for parametric and vine copulas models and functional principal component analysis (fPCA) -- an extension of principal component analysis where data consist of functions instead of vectors [@Ramsay_2005] -- to allow for a wide range of data and dependent structures to be modelled. For efficiency, algorithms are implemented in NumPy [@Harris_2020] and SciPy [@SciPy_2020] for Gaussian (parametric) copula and fPCA classes and rely on the C++ library vinecopulib  [@Nagler_2020a] through pyvinecopulib's  [@Nagler_2020] bindings for fast computation of vines.
+Here we introduce Synthia, an open-source multidimensional synthetic data generator written in Python for xarray's [@Hoyer_2017] labelled arrays and datasets with support for parametric and vine copulas models and functional principal component analysis (fPCA) -- an extension of principal component analysis where data consist of functions instead of vectors [@Ramsay_2005] -- to allow for a wide range of data and dependent structures to be modelled. For efficiency, algorithms are implemented in NumPy [@Harris_2020] and SciPy [@SciPy_2020] for Gaussian (parametric) copula and fPCA classes and rely on the C++ library vinecopulib [@Nagler_2020a] through pyvinecopulib's [@Nagler_2020] bindings for fast computation of vines.
 
[…]
