+++ paper/paper.bib
@@ -96,20 +96,22 @@
   doi = {10.5194/gmd-14-5205-2021}
 }
 
-@misc{Nagler_2020,
-  title = {Pyvinecopulib},
-  author = {Nagler, Thomas and Vatter, Thibault},
-  year = {2020},
-  howpublished = {\url{https://doi.org/10.5281/zenodo.4288292}},
-  keywords = {copula,dependence,pair copula construction,simulation,vine}
+@software{Nagler_2020a,
+  author       = {Nagler, Thomas and Vatter, Thibault},
+  title        = {vinecopulib},
+  year         = 2020,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.4287554},
+  url          = {https://doi.org/10.5281/zenodo.4287554}
 }
 
-@misc{Nagler_2020a,
-  title = {Vinecopulib},
-  author = {Nagler, Thomas and Vatter, Thibault},
-  year = {2020},
-  howpublished = {\url{https://doi.org/10.5281/zenodo.4287554}},
-  keywords = {copula,dependence,pair copula construction,simulation,vine}
+@software{Nagler_2020b,
+  author       = {Nagler, Thomas and Vatter, Thibault},
+  title        = {pyvinecopulib},
+  year         = 2020,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.4288292},
+  url          = {https://doi.org/10.5281/zenodo.4288292}
 }
 
 @article{Nowok_2016,
