### 2021-08-18 #2863 @dmey
@khinsen and @oliviaguest thanks for waiting -- I have now updated references and paper. @khinsen a couple of things:
- (a) In the rendered PDF `(Nagler & Vatter, 2020b)` comes before `(Nagler & Vatter, 2020a)` therefore it should be marked with `a` instead of `b` -- would you know how to change this?
- (b) I have updated the main heading from 'Summary' to 'Summary and Statement of Need' as instructed however I noticed that this is not always enforced by editors -- if this is not a strong requirement any longer would it be possible to switch back to just 'Summary'?
