### 2021-08-20 #2863 @dmey
Sorry @khinsen meant to contact @danielskatz instead! @danielskatz can you please let me know about https://github.com/openjournals/joss-reviews/issues/2863#issuecomment-901301341?
