### 2021-08-20 #2863 @danielskatz
Re a, this is because the references are sorted by author-year-title, not by appearance, so I don't think you can change this.

Re b, no. We try to do this in all paper (since 1-2 years ago), but sometimes we may fail. But it is our policy to do this, and we will do it here.
