### 2021-09-01 c96ed03
update dev docs to always install full version with pyvinecopulib
