### 2021-09-01 3d26d6d
update docs and CHANGELOG.txt for version 1.1.0.
