+++ CHANGELOG.txt
@@ -1,3 +1,8 @@
+1.1.0
+- Pin to specific pyvinecopulib version to avoid issues between versions.
+- Add CI tests for Python 3.9 (#17).
+- Minor doc improvements.
+
 1.0.0
 - Add JOSS summary paper (#26).
 - Improve docs and tutorials (#14, #13, #18, ...).
