### 2021-09-01 6082200
Minor doc and CI updates + pin pyvinecopulib version (#27)
