+++ .github/workflows/ci.yml
@@ -11,7 +11,7 @@ jobs:
       matrix:
         os: ['ubuntu-latest', 'macos-latest', 'windows-latest']
         # when changing version, also change setup.py
-        python-version: ['3.8']
+        python-version: ['3.8', '3.9']
     steps:
     - uses: actions/checkout@v2
     - uses: conda-incubator/setup-miniconda@v2
@@ -31,11 +31,11 @@ jobs:
       shell: bash -l {0}
       run: pytest tests/
 
-    - name: Run tests with pyvinecopulib
+    - name: Run tests with pyvinecopulib==0.5.5
       shell: bash -l {0}
       run: |
         set -ex
-        pip install pyvinecopulib
+        pip install pyvinecopulib==0.5.5
         pytest tests/
 
   gh-pages:
@@ -49,6 +49,9 @@ jobs:
         python-version: 3.8
         activate-environment: synthia
         environment-file: environment.yml
+
+    - name: Install pandoc
+      run: sudo apt-get install pandoc
 
     - name: Install Synthia (dev env)
       # 'shell' required to activate environment.
