### 2021-09-01 #2863 @dmey
@danielskatz and @oliviaguest would you have an update about this submission? Please let me know if you are waiting for further changes from my side. If not please can you use https://github.com/openjournals/joss-reviews/issues/2863#issuecomment-901292885 as final version for the summary-paper and version 1.1.0 with corresponding Zenodo archive https://zenodo.org/record/5358432 (currently set to an older version).
