### 2021-09-23 #2863 @dmey
@oliviaguest and @kthyng I am happy with the final proof, thank you. Re DOI showing as invalid this is correct but it has not yet been accepted (and deposited). The DOI will not change so I can (a) leave it as is, (b) update the reference to include 'under review', or (c) wait for the DOI to be deposited -- this should take a few weeks. Happy with any of those options.
